Added Accel_Inertial and main updates
Added Accel_Inertial function and main updates to interact with it. Created a for loop to test it (this is not where the for loop will be for overall testing, just testing of this function). The incoming gyro and accel_body are set to ones and the star_tracker set to 0's every 10 Hz)